Rename exportdata related modules
Rename dmaws.syncdata to dmaws.rds as it's mostly about interacting with
RDS.
Rename dmaws.commands.syncdata to dmaws.commands.exportdata as it
doesn't handle importing into other databases just exporting.
